@@ -1,28 +1,10 @@
 {
-  "name": "heron-handoff",
-  "description": "A developer-handoff tool for Figma.",
-  "keywords": [
-    "figma",
-    "design",
-    "developer-handoff",
-    "handoff",
-    "specs"
-  ],
-  "bugs": {
-    "url": "https://github.com/leadream/heron-handoff/issues"
-  },
-  "homepage": "https://github.com/leadream/heron-handoff#readme",
-  "author": "leadream",
-  "repository": "https://github.com/leadream/heron-handoff.git",
-  "version": "1.2.1",
-  "license": "AGPL-3.0",
+  "name": "f2c-handoff",
+  "version": "0.0.1",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
+  "files": ["dist"],
   "dependencies": {
-    "@empjs/cli": "^3.0.1",
     "classnames": "^2.2.6",
     "color": "^3.1.2",
     "file-saver": "^2.0.2",
@@ -32,44 +14,25 @@
     "prismjs": "^1.23.0",
     "rc-input-number": "^4.5.3",
     "rc-tooltip": "^4.0.0-alpha.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "18",
+    "react-dom": "18",
     "react-feather": "^2.0.3",
     "react-i18next": "^11.3.1"
   },
-  "peerDependencies": {
-    "react": "^16.8.6 || ^17",
-    "react-dom": "^16.8.6 || ^17"
-  },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "dev": "emp dev",
+    "build": "emp build",
+    "build:watch": "emp build --watch --serve",
+    "start": "emp serve",
+    "stat": "emp build --analyze",
+    "lint": "biome check . --fix",
     "watch:npm": "rollup -c -w",
     "build:npm": "rollup -c --environment NODE_ENV:production"
   },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
   "devDependencies": {
-    "@babel/cli": "^7.13.14",
-    "@babel/core": "^7.13.15",
-    "@babel/plugin-proposal-class-properties": "^7.13.0",
-    "@babel/preset-env": "^7.13.15",
-    "@babel/preset-react": "^7.13.13",
+    "@empjs/cli": "^3.0.1",
+    "@empjs/biome-config": "^0.4.2",
+    "@empjs/plugin-react": "^3.0.0",
     "@jackfranklin/rollup-plugin-markdown": "^0.3.0",
     "@rollup/plugin-alias": "^3.1.2",
     "@rollup/plugin-commonjs": "^18.0.0",
@@ -77,14 +40,12 @@
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^11.2.1",
     "@rollup/plugin-replace": "^2.4.2",
-    "react-scripts": "4.0.3",
     "rollup": "^2.45.2",
     "rollup-plugin-babel": "^4.4.0",
     "rollup-plugin-copy": "^3.4.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-scss": "^2.6.1",
     "rollup-plugin-uglify": "^6.0.4",
-    "rollup-plugin-visualizer": "^5.3.6",
-    "sass": "^1.56.1"
+    "rollup-plugin-visualizer": "^5.3.6"
   }
 }
