@@ -1,18 +1,19 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import { I18nextProvider } from 'react-i18next'
-import i18n from './i18n'
-import { withGlobalContextProvider } from 'contexts/GlobalContext'
+import {withGlobalContextProvider} from 'contexts/GlobalContext'
+
+import {createRoot} from 'react-dom/client'
+import {I18nextProvider} from 'react-i18next'
 import App from './App'
+import i18n from './i18n'
 import * as serviceWorker from './serviceWorker'
 
 const ContextedApp = withGlobalContextProvider(App)
 
-ReactDOM.render(
+const dom = document.getElementById('root')
+const root = createRoot(dom)
+root.render(
   <I18nextProvider i18n={i18n}>
     <ContextedApp />
   </I18nextProvider>,
-  document.getElementById('root')
 )
 
 // If you want your app to work offline and load faster, you can change
