@@ -1,25 +1,23 @@
-import path from 'path'
-import babel from 'rollup-plugin-babel';
-import commonjs from '@rollup/plugin-commonjs';
-import external from 'rollup-plugin-peer-deps-external';
-import scss from 'rollup-plugin-scss';
-import resolve from '@rollup/plugin-node-resolve';
-import image from '@rollup/plugin-image'
+import path from 'node:path'
 import markdown from '@jackfranklin/rollup-plugin-markdown'
-import visualizer from 'rollup-plugin-visualizer';
-import alias from '@rollup/plugin-alias';
-import json from '@rollup/plugin-json';
-import copy from 'rollup-plugin-copy';
-import { uglify } from 'rollup-plugin-uglify';
-import replace from '@rollup/plugin-replace';
-import pkg from './package.json';
+import alias from '@rollup/plugin-alias'
+import commonjs from '@rollup/plugin-commonjs'
+import image from '@rollup/plugin-image'
+import json from '@rollup/plugin-json'
+import resolve from '@rollup/plugin-node-resolve'
+import replace from '@rollup/plugin-replace'
+import babel from 'rollup-plugin-babel'
+import copy from 'rollup-plugin-copy'
+import external from 'rollup-plugin-peer-deps-external'
+import scss from 'rollup-plugin-scss'
+import {uglify} from 'rollup-plugin-uglify'
+import visualizer from 'rollup-plugin-visualizer'
+import pkg from './package.json'
 
 const mode = process.env.NODE_ENV || 'development'
-const isProduction = mode==='production'
+const isProduction = mode === 'production'
 
-const resolveFile = function(filePath) {
-  return path.join(__dirname, filePath)
-}
+const resolveFile = filePath => path.join(__dirname, filePath)
 
 const config = {
   input: './src/lib/index.js',
@@ -28,8 +26,8 @@ const config = {
       sourcemap: isProduction ? false : 'inline',
       file: pkg.main,
       format: 'cjs',
-      exports: 'named'
-    }
+      exports: 'named',
+    },
   ],
   external: ['react', 'react-dom'],
   plugins: [
@@ -40,25 +38,23 @@ const config = {
         utils: resolveFile('src/utils'),
         contexts: resolveFile('src/contexts'),
         page: resolveFile('src/page'),
-        api: resolveFile('src/api')
-      }
+        api: resolveFile('src/api'),
+      },
     }),
     external(),
     json(),
     copy({
-      targets: [
-        { src: resolveFile('src/lib/index.d.ts'), dest: resolveFile('dist') }
-      ]
+      targets: [{src: resolveFile('src/lib/index.d.ts'), dest: resolveFile('dist')}],
     }),
     scss({
-      outputStyle: isProduction ? 'compressed' : 'compact'
+      outputStyle: isProduction ? 'compressed' : 'compact',
     }),
     babel({
-      exclude: 'node_modules/**'
+      exclude: 'node_modules/**',
     }),
     resolve({
       preferBuiltins: true,
-      mainFields: ['browser']
+      mainFields: ['browser'],
     }),
     commonjs(),
     image(),
@@ -67,12 +63,12 @@ const config = {
       preventAssignment: true,
       values: {
         'process.env.IS_MODULE': JSON.stringify(true),
-        'process.env.NODE_ENV': JSON.stringify(mode)
-      }
+        'process.env.NODE_ENV': JSON.stringify(mode),
+      },
     }),
-    visualizer()
-  ]
-};
+    visualizer(),
+  ],
+}
 
 if (isProduction) {
   config.plugins.push(uglify())
