@@ -2,9 +2,9 @@
 export const getMockFile = () =>
   fetch(`${process.env.PUBLIC_URL}/data/file.json`)
     .then(response => response.json())
-    .catch(function (error) {
+    .catch(error => {
       console.dir(error)
-      return { err: error }
+      return {err: error}
     })
 
 // get buffer data
@@ -13,7 +13,7 @@ export const getBlobData = url =>
     .then(response => response.blob())
     .catch(error => {
       console.dir(error)
-      return { err: error }
+      return {err: error}
     })
 
 // get buffer data
@@ -22,6 +22,6 @@ export const getBufferData = url => {
     .then(response => response.arrayBuffer())
     .catch(error => {
       console.dir(error)
-      return { err: error }
+      return {err: error}
     })
 }
