@@ -1,55 +1,127 @@
 import React from 'react'
 
-export const SmallThin = ({size=24, className}) =>
-  <svg className={className} width={size} height={size} viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-    <path d="M7.65604 12.336L7.99804 11.265C8.24104 10.518 8.45704 9.816 8.67304 9.033H8.70904C8.93404 9.807 9.14104 10.518 9.39304 11.265L9.73504 12.336H7.65604ZM10.581 15H11.373L9.11404 8.412H8.29504L6.04504 15H6.79204L7.46704 12.921H9.92404L10.581 15Z"/>
-    <path d="M14.7803 17.259C16.2563 17.259 17.2013 16.476 17.2013 15.594C17.2013 14.793 16.6433 14.442 15.5273 14.442H14.5553C13.8893 14.442 13.6913 14.208 13.6913 13.893C13.6913 13.614 13.8353 13.434 14.0153 13.281C14.2403 13.398 14.5193 13.461 14.7623 13.461C15.7433 13.461 16.5173 12.795 16.5173 11.76C16.5173 11.319 16.3373 10.941 16.0943 10.707H17.1203V10.14H15.4373C15.2663 10.077 15.0323 10.014 14.7623 10.014C13.7813 10.014 12.9623 10.698 12.9623 11.751C12.9623 12.327 13.2773 12.804 13.5923 13.056V13.092C13.3493 13.272 13.0523 13.587 13.0523 13.992C13.0523 14.379 13.2413 14.64 13.4933 14.793V14.829C13.0433 15.126 12.7733 15.531 12.7733 15.945C12.7733 16.782 13.5833 17.259 14.7803 17.259ZM14.7623 12.957C14.1773 12.957 13.6733 12.48 13.6733 11.751C13.6733 11.004 14.1593 10.563 14.7623 10.563C15.3563 10.563 15.8513 11.013 15.8513 11.751C15.8513 12.48 15.3473 12.957 14.7623 12.957ZM14.8793 16.746C13.9613 16.746 13.4213 16.395 13.4213 15.855C13.4213 15.558 13.5833 15.243 13.9523 14.982C14.1773 15.036 14.4203 15.063 14.5733 15.063H15.4643C16.1303 15.063 16.4813 15.225 16.4813 15.702C16.4813 16.224 15.8513 16.746 14.8793 16.746Z"/>
+export const SmallThin = ({size = 24, className}) => (
+  <svg
+    className={className}
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M7.65604 12.336L7.99804 11.265C8.24104 10.518 8.45704 9.816 8.67304 9.033H8.70904C8.93404 9.807 9.14104 10.518 9.39304 11.265L9.73504 12.336H7.65604ZM10.581 15H11.373L9.11404 8.412H8.29504L6.04504 15H6.79204L7.46704 12.921H9.92404L10.581 15Z" />
+    <path d="M14.7803 17.259C16.2563 17.259 17.2013 16.476 17.2013 15.594C17.2013 14.793 16.6433 14.442 15.5273 14.442H14.5553C13.8893 14.442 13.6913 14.208 13.6913 13.893C13.6913 13.614 13.8353 13.434 14.0153 13.281C14.2403 13.398 14.5193 13.461 14.7623 13.461C15.7433 13.461 16.5173 12.795 16.5173 11.76C16.5173 11.319 16.3373 10.941 16.0943 10.707H17.1203V10.14H15.4373C15.2663 10.077 15.0323 10.014 14.7623 10.014C13.7813 10.014 12.9623 10.698 12.9623 11.751C12.9623 12.327 13.2773 12.804 13.5923 13.056V13.092C13.3493 13.272 13.0523 13.587 13.0523 13.992C13.0523 14.379 13.2413 14.64 13.4933 14.793V14.829C13.0433 15.126 12.7733 15.531 12.7733 15.945C12.7733 16.782 13.5833 17.259 14.7803 17.259ZM14.7623 12.957C14.1773 12.957 13.6733 12.48 13.6733 11.751C13.6733 11.004 14.1593 10.563 14.7623 10.563C15.3563 10.563 15.8513 11.013 15.8513 11.751C15.8513 12.48 15.3473 12.957 14.7623 12.957ZM14.8793 16.746C13.9613 16.746 13.4213 16.395 13.4213 15.855C13.4213 15.558 13.5833 15.243 13.9523 14.982C14.1773 15.036 14.4203 15.063 14.5733 15.063H15.4643C16.1303 15.063 16.4813 15.225 16.4813 15.702C16.4813 16.224 15.8513 16.746 14.8793 16.746Z" />
   </svg>
+)
 
-export const NormalThin = ({size=24, className}) =>
-  <svg className={className} width={size} height={size} viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-    <path d="M6.20806 13.448L6.66406 12.02C6.98806 11.024 7.27606 10.088 7.56406 9.044H7.61206C7.91206 10.076 8.18806 11.024 8.52406 12.02L8.98006 13.448H6.20806ZM10.1081 17H11.1641L8.15206 8.216H7.06006L4.06006 17H5.05606L5.95606 14.228H9.23206L10.1081 17Z"/>
-    <path d="M15.7071 20.012C17.6751 20.012 18.9351 18.968 18.9351 17.792C18.9351 16.724 18.1911 16.256 16.7031 16.256H15.4071C14.5191 16.256 14.2551 15.944 14.2551 15.524C14.2551 15.152 14.4471 14.912 14.6871 14.708C14.9871 14.864 15.3591 14.948 15.6831 14.948C16.9911 14.948 18.0231 14.06 18.0231 12.68C18.0231 12.092 17.7831 11.588 17.4591 11.276H18.8271V10.52H16.5831C16.3551 10.436 16.0431 10.352 15.6831 10.352C14.3751 10.352 13.2831 11.264 13.2831 12.668C13.2831 13.436 13.7031 14.072 14.1231 14.408V14.456C13.7991 14.696 13.4031 15.116 13.4031 15.656C13.4031 16.172 13.6551 16.52 13.9911 16.724V16.772C13.3911 17.168 13.0311 17.708 13.0311 18.26C13.0311 19.376 14.1111 20.012 15.7071 20.012ZM15.6831 14.276C14.9031 14.276 14.2311 13.64 14.2311 12.668C14.2311 11.672 14.8791 11.084 15.6831 11.084C16.4751 11.084 17.1351 11.684 17.1351 12.668C17.1351 13.64 16.4631 14.276 15.6831 14.276ZM15.8391 19.328C14.6151 19.328 13.8951 18.86 13.8951 18.14C13.8951 17.744 14.1111 17.324 14.6031 16.976C14.9031 17.048 15.2271 17.084 15.4311 17.084H16.6191C17.5071 17.084 17.9751 17.3 17.9751 17.936C17.9751 18.632 17.1351 19.328 15.8391 19.328Z"/>
+export const NormalThin = ({size = 24, className}) => (
+  <svg
+    className={className}
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M6.20806 13.448L6.66406 12.02C6.98806 11.024 7.27606 10.088 7.56406 9.044H7.61206C7.91206 10.076 8.18806 11.024 8.52406 12.02L8.98006 13.448H6.20806ZM10.1081 17H11.1641L8.15206 8.216H7.06006L4.06006 17H5.05606L5.95606 14.228H9.23206L10.1081 17Z" />
+    <path d="M15.7071 20.012C17.6751 20.012 18.9351 18.968 18.9351 17.792C18.9351 16.724 18.1911 16.256 16.7031 16.256H15.4071C14.5191 16.256 14.2551 15.944 14.2551 15.524C14.2551 15.152 14.4471 14.912 14.6871 14.708C14.9871 14.864 15.3591 14.948 15.6831 14.948C16.9911 14.948 18.0231 14.06 18.0231 12.68C18.0231 12.092 17.7831 11.588 17.4591 11.276H18.8271V10.52H16.5831C16.3551 10.436 16.0431 10.352 15.6831 10.352C14.3751 10.352 13.2831 11.264 13.2831 12.668C13.2831 13.436 13.7031 14.072 14.1231 14.408V14.456C13.7991 14.696 13.4031 15.116 13.4031 15.656C13.4031 16.172 13.6551 16.52 13.9911 16.724V16.772C13.3911 17.168 13.0311 17.708 13.0311 18.26C13.0311 19.376 14.1111 20.012 15.7071 20.012ZM15.6831 14.276C14.9031 14.276 14.2311 13.64 14.2311 12.668C14.2311 11.672 14.8791 11.084 15.6831 11.084C16.4751 11.084 17.1351 11.684 17.1351 12.668C17.1351 13.64 16.4631 14.276 15.6831 14.276ZM15.8391 19.328C14.6151 19.328 13.8951 18.86 13.8951 18.14C13.8951 17.744 14.1111 17.324 14.6031 16.976C14.9031 17.048 15.2271 17.084 15.4311 17.084H16.6191C17.5071 17.084 17.9751 17.3 17.9751 17.936C17.9751 18.632 17.1351 19.328 15.8391 19.328Z" />
   </svg>
+)
 
-export const LargeThin = ({size=24, className}) =>
-  <svg className={className} width={size} height={size} viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-    <path d="M4.75995 12.56L5.32995 10.775C5.73495 9.53 6.09495 8.36 6.45495 7.055H6.51495C6.88995 8.34501 7.23495 9.53 7.65495 10.775L8.22495 12.56H4.75995ZM9.63495 17H10.955L7.18995 6.02H5.82495L2.07495 17H3.31995L4.44495 13.535H8.53995L9.63495 17Z"/>
-    <path d="M16.6337 20.765C19.0937 20.765 20.6687 19.46 20.6687 17.99C20.6687 16.655 19.7387 16.07 17.8787 16.07H16.2587C15.1487 16.07 14.8187 15.68 14.8187 15.155C14.8187 14.69 15.0587 14.39 15.3587 14.135C15.7337 14.33 16.1987 14.435 16.6037 14.435C18.2387 14.435 19.5287 13.325 19.5287 11.6C19.5287 10.865 19.2287 10.235 18.8237 9.845H20.5337V8.9H17.7287C17.4437 8.795 17.0537 8.69 16.6037 8.69C14.9687 8.69 13.6037 9.83 13.6037 11.585C13.6037 12.545 14.1287 13.34 14.6537 13.76V13.82C14.2487 14.12 13.7537 14.645 13.7537 15.32C13.7537 15.965 14.0687 16.4 14.4887 16.655V16.715C13.7387 17.21 13.2887 17.885 13.2887 18.575C13.2887 19.97 14.6387 20.765 16.6337 20.765ZM16.6037 13.595C15.6287 13.595 14.7887 12.8 14.7887 11.585C14.7887 10.34 15.5987 9.605 16.6037 9.605C17.5937 9.605 18.4187 10.355 18.4187 11.585C18.4187 12.8 17.5787 13.595 16.6037 13.595ZM16.7987 19.91C15.2687 19.91 14.3687 19.325 14.3687 18.425C14.3687 17.93 14.6387 17.405 15.2537 16.97C15.6287 17.06 16.0337 17.105 16.2887 17.105H17.7737C18.8837 17.105 19.4687 17.375 19.4687 18.17C19.4687 19.04 18.4187 19.91 16.7987 19.91Z"/>
+export const LargeThin = ({size = 24, className}) => (
+  <svg
+    className={className}
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M4.75995 12.56L5.32995 10.775C5.73495 9.53 6.09495 8.36 6.45495 7.055H6.51495C6.88995 8.34501 7.23495 9.53 7.65495 10.775L8.22495 12.56H4.75995ZM9.63495 17H10.955L7.18995 6.02H5.82495L2.07495 17H3.31995L4.44495 13.535H8.53995L9.63495 17Z" />
+    <path d="M16.6337 20.765C19.0937 20.765 20.6687 19.46 20.6687 17.99C20.6687 16.655 19.7387 16.07 17.8787 16.07H16.2587C15.1487 16.07 14.8187 15.68 14.8187 15.155C14.8187 14.69 15.0587 14.39 15.3587 14.135C15.7337 14.33 16.1987 14.435 16.6037 14.435C18.2387 14.435 19.5287 13.325 19.5287 11.6C19.5287 10.865 19.2287 10.235 18.8237 9.845H20.5337V8.9H17.7287C17.4437 8.795 17.0537 8.69 16.6037 8.69C14.9687 8.69 13.6037 9.83 13.6037 11.585C13.6037 12.545 14.1287 13.34 14.6537 13.76V13.82C14.2487 14.12 13.7537 14.645 13.7537 15.32C13.7537 15.965 14.0687 16.4 14.4887 16.655V16.715C13.7387 17.21 13.2887 17.885 13.2887 18.575C13.2887 19.97 14.6387 20.765 16.6337 20.765ZM16.6037 13.595C15.6287 13.595 14.7887 12.8 14.7887 11.585C14.7887 10.34 15.5987 9.605 16.6037 9.605C17.5937 9.605 18.4187 10.355 18.4187 11.585C18.4187 12.8 17.5787 13.595 16.6037 13.595ZM16.7987 19.91C15.2687 19.91 14.3687 19.325 14.3687 18.425C14.3687 17.93 14.6387 17.405 15.2537 16.97C15.6287 17.06 16.0337 17.105 16.2887 17.105H17.7737C18.8837 17.105 19.4687 17.375 19.4687 18.17C19.4687 19.04 18.4187 19.91 16.7987 19.91Z" />
   </svg>
+)
 
-export const SmallRegular = ({size=24, className}) =>
-  <svg className={className} width={size} height={size} viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-    <path d="M7.872 12.291L8.142 11.4C8.358 10.68 8.565 9.951 8.754 9.195H8.79C8.997 9.942 9.195 10.68 9.42 11.4L9.681 12.291H7.872ZM10.491 15H11.598L9.402 8.376H8.196L6 15H7.062L7.629 13.11H9.933L10.491 15Z"/>
-    <path d="M14.9736 17.223C16.5576 17.223 17.5656 16.449 17.5656 15.495C17.5656 14.649 16.9536 14.289 15.7836 14.289H14.8746C14.2446 14.289 14.0466 14.1 14.0466 13.803C14.0466 13.569 14.1546 13.425 14.3166 13.281C14.5326 13.38 14.7846 13.434 15.0096 13.434C16.0446 13.434 16.8636 12.813 16.8636 11.733C16.8636 11.355 16.7286 11.022 16.5396 10.824H17.4846V10.05H15.7296C15.5406 9.978 15.2886 9.924 15.0096 9.924C13.9836 9.924 13.0926 10.581 13.0926 11.706C13.0926 12.291 13.4076 12.759 13.7406 13.02V13.056C13.4706 13.254 13.1916 13.587 13.1916 13.974C13.1916 14.379 13.3896 14.64 13.6416 14.793V14.838C13.1826 15.117 12.9216 15.504 12.9216 15.918C12.9216 16.782 13.7856 17.223 14.9736 17.223ZM15.0096 12.768C14.4966 12.768 14.0736 12.363 14.0736 11.706C14.0736 11.031 14.4876 10.653 15.0096 10.653C15.5316 10.653 15.9456 11.04 15.9456 11.706C15.9456 12.363 15.5226 12.768 15.0096 12.768ZM15.1266 16.539C14.3076 16.539 13.8036 16.251 13.8036 15.765C13.8036 15.513 13.9296 15.261 14.2266 15.045C14.4336 15.09 14.6586 15.117 14.8926 15.117H15.6396C16.2426 15.117 16.5666 15.243 16.5666 15.657C16.5666 16.116 15.9906 16.539 15.1266 16.539Z"/>
+export const SmallRegular = ({size = 24, className}) => (
+  <svg
+    className={className}
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M7.872 12.291L8.142 11.4C8.358 10.68 8.565 9.951 8.754 9.195H8.79C8.997 9.942 9.195 10.68 9.42 11.4L9.681 12.291H7.872ZM10.491 15H11.598L9.402 8.376H8.196L6 15H7.062L7.629 13.11H9.933L10.491 15Z" />
+    <path d="M14.9736 17.223C16.5576 17.223 17.5656 16.449 17.5656 15.495C17.5656 14.649 16.9536 14.289 15.7836 14.289H14.8746C14.2446 14.289 14.0466 14.1 14.0466 13.803C14.0466 13.569 14.1546 13.425 14.3166 13.281C14.5326 13.38 14.7846 13.434 15.0096 13.434C16.0446 13.434 16.8636 12.813 16.8636 11.733C16.8636 11.355 16.7286 11.022 16.5396 10.824H17.4846V10.05H15.7296C15.5406 9.978 15.2886 9.924 15.0096 9.924C13.9836 9.924 13.0926 10.581 13.0926 11.706C13.0926 12.291 13.4076 12.759 13.7406 13.02V13.056C13.4706 13.254 13.1916 13.587 13.1916 13.974C13.1916 14.379 13.3896 14.64 13.6416 14.793V14.838C13.1826 15.117 12.9216 15.504 12.9216 15.918C12.9216 16.782 13.7856 17.223 14.9736 17.223ZM15.0096 12.768C14.4966 12.768 14.0736 12.363 14.0736 11.706C14.0736 11.031 14.4876 10.653 15.0096 10.653C15.5316 10.653 15.9456 11.04 15.9456 11.706C15.9456 12.363 15.5226 12.768 15.0096 12.768ZM15.1266 16.539C14.3076 16.539 13.8036 16.251 13.8036 15.765C13.8036 15.513 13.9296 15.261 14.2266 15.045C14.4336 15.09 14.6586 15.117 14.8926 15.117H15.6396C16.2426 15.117 16.5666 15.243 16.5666 15.657C16.5666 16.116 15.9906 16.539 15.1266 16.539Z" />
   </svg>
+)
 
-export const NormalRegular = ({size=24, className}) =>
-  <svg className={className} width={size} height={size} viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-    <path d="M6.496 13.388L6.856 12.2C7.144 11.24 7.42 10.268 7.672 9.26H7.72C7.996 10.256 8.26 11.24 8.56 12.2L8.908 13.388H6.496ZM9.988 17H11.464L8.536 8.168H6.928L4 17H5.416L6.172 14.48H9.244L9.988 17Z"/>
-    <path d="M15.9648 19.964C18.0768 19.964 19.4208 18.932 19.4208 17.66C19.4208 16.532 18.6048 16.052 17.0448 16.052H15.8328C14.9928 16.052 14.7288 15.8 14.7288 15.404C14.7288 15.092 14.8728 14.9 15.0888 14.708C15.3768 14.84 15.7128 14.912 16.0128 14.912C17.3928 14.912 18.4848 14.084 18.4848 12.644C18.4848 12.14 18.3048 11.696 18.0528 11.432H19.3128V10.4H16.9728C16.7208 10.304 16.3848 10.232 16.0128 10.232C14.6448 10.232 13.4568 11.108 13.4568 12.608C13.4568 13.388 13.8768 14.012 14.3208 14.36V14.408C13.9608 14.672 13.5888 15.116 13.5888 15.632C13.5888 16.172 13.8528 16.52 14.1888 16.724V16.784C13.5768 17.156 13.2288 17.672 13.2288 18.224C13.2288 19.376 14.3808 19.964 15.9648 19.964ZM16.0128 14.024C15.3288 14.024 14.7648 13.484 14.7648 12.608C14.7648 11.708 15.3168 11.204 16.0128 11.204C16.7088 11.204 17.2608 11.72 17.2608 12.608C17.2608 13.484 16.6968 14.024 16.0128 14.024ZM16.1688 19.052C15.0768 19.052 14.4048 18.668 14.4048 18.02C14.4048 17.684 14.5728 17.348 14.9688 17.06C15.2448 17.12 15.5448 17.156 15.8568 17.156H16.8528C17.6568 17.156 18.0888 17.324 18.0888 17.876C18.0888 18.488 17.3208 19.052 16.1688 19.052Z"/>
+export const NormalRegular = ({size = 24, className}) => (
+  <svg
+    className={className}
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M6.496 13.388L6.856 12.2C7.144 11.24 7.42 10.268 7.672 9.26H7.72C7.996 10.256 8.26 11.24 8.56 12.2L8.908 13.388H6.496ZM9.988 17H11.464L8.536 8.168H6.928L4 17H5.416L6.172 14.48H9.244L9.988 17Z" />
+    <path d="M15.9648 19.964C18.0768 19.964 19.4208 18.932 19.4208 17.66C19.4208 16.532 18.6048 16.052 17.0448 16.052H15.8328C14.9928 16.052 14.7288 15.8 14.7288 15.404C14.7288 15.092 14.8728 14.9 15.0888 14.708C15.3768 14.84 15.7128 14.912 16.0128 14.912C17.3928 14.912 18.4848 14.084 18.4848 12.644C18.4848 12.14 18.3048 11.696 18.0528 11.432H19.3128V10.4H16.9728C16.7208 10.304 16.3848 10.232 16.0128 10.232C14.6448 10.232 13.4568 11.108 13.4568 12.608C13.4568 13.388 13.8768 14.012 14.3208 14.36V14.408C13.9608 14.672 13.5888 15.116 13.5888 15.632C13.5888 16.172 13.8528 16.52 14.1888 16.724V16.784C13.5768 17.156 13.2288 17.672 13.2288 18.224C13.2288 19.376 14.3808 19.964 15.9648 19.964ZM16.0128 14.024C15.3288 14.024 14.7648 13.484 14.7648 12.608C14.7648 11.708 15.3168 11.204 16.0128 11.204C16.7088 11.204 17.2608 11.72 17.2608 12.608C17.2608 13.484 16.6968 14.024 16.0128 14.024ZM16.1688 19.052C15.0768 19.052 14.4048 18.668 14.4048 18.02C14.4048 17.684 14.5728 17.348 14.9688 17.06C15.2448 17.12 15.5448 17.156 15.8568 17.156H16.8528C17.6568 17.156 18.0888 17.324 18.0888 17.876C18.0888 18.488 17.3208 19.052 16.1688 19.052Z" />
   </svg>
+)
 
-export const LargeRegular = ({size=24, className}) =>
-  <svg className={className} width={size} height={size} viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-    <path d="M5.12 12.485L5.57 11C5.93 9.8 6.275 8.585 6.59 7.325H6.65C6.995 8.57 7.325 9.8 7.7 11L8.135 12.485H5.12ZM9.485 17H11.33L7.67 5.96H5.66L2 17H3.77L4.715 13.85H8.555L9.485 17Z"/>
-    <path d="M16.9561 20.705C19.5961 20.705 21.2761 19.415 21.2761 17.825C21.2761 16.415 20.2561 15.815 18.3061 15.815H16.7911C15.7411 15.815 15.4111 15.5 15.4111 15.005C15.4111 14.615 15.5911 14.375 15.8611 14.135C16.2211 14.3 16.6411 14.39 17.0161 14.39C18.7411 14.39 20.1061 13.355 20.1061 11.555C20.1061 10.925 19.8811 10.37 19.5661 10.04H21.1411V8.75H18.2161C17.9011 8.63 17.4811 8.54 17.0161 8.54C15.3061 8.54 13.8211 9.635 13.8211 11.51C13.8211 12.485 14.3461 13.265 14.9011 13.7V13.76C14.4511 14.09 13.9861 14.645 13.9861 15.29C13.9861 15.965 14.3161 16.4 14.7361 16.655V16.73C13.9711 17.195 13.5361 17.84 13.5361 18.53C13.5361 19.97 14.9761 20.705 16.9561 20.705ZM17.0161 13.28C16.1611 13.28 15.4561 12.605 15.4561 11.51C15.4561 10.385 16.1461 9.755 17.0161 9.755C17.8861 9.755 18.5761 10.4 18.5761 11.51C18.5761 12.605 17.8711 13.28 17.0161 13.28ZM17.2111 19.565C15.8461 19.565 15.0061 19.085 15.0061 18.275C15.0061 17.855 15.2161 17.435 15.7111 17.075C16.0561 17.15 16.4311 17.195 16.8211 17.195H18.0661C19.0711 17.195 19.6111 17.405 19.6111 18.095C19.6111 18.86 18.6511 19.565 17.2111 19.565Z"/>
+export const LargeRegular = ({size = 24, className}) => (
+  <svg
+    className={className}
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M5.12 12.485L5.57 11C5.93 9.8 6.275 8.585 6.59 7.325H6.65C6.995 8.57 7.325 9.8 7.7 11L8.135 12.485H5.12ZM9.485 17H11.33L7.67 5.96H5.66L2 17H3.77L4.715 13.85H8.555L9.485 17Z" />
+    <path d="M16.9561 20.705C19.5961 20.705 21.2761 19.415 21.2761 17.825C21.2761 16.415 20.2561 15.815 18.3061 15.815H16.7911C15.7411 15.815 15.4111 15.5 15.4111 15.005C15.4111 14.615 15.5911 14.375 15.8611 14.135C16.2211 14.3 16.6411 14.39 17.0161 14.39C18.7411 14.39 20.1061 13.355 20.1061 11.555C20.1061 10.925 19.8811 10.37 19.5661 10.04H21.1411V8.75H18.2161C17.9011 8.63 17.4811 8.54 17.0161 8.54C15.3061 8.54 13.8211 9.635 13.8211 11.51C13.8211 12.485 14.3461 13.265 14.9011 13.7V13.76C14.4511 14.09 13.9861 14.645 13.9861 15.29C13.9861 15.965 14.3161 16.4 14.7361 16.655V16.73C13.9711 17.195 13.5361 17.84 13.5361 18.53C13.5361 19.97 14.9761 20.705 16.9561 20.705ZM17.0161 13.28C16.1611 13.28 15.4561 12.605 15.4561 11.51C15.4561 10.385 16.1461 9.755 17.0161 9.755C17.8861 9.755 18.5761 10.4 18.5761 11.51C18.5761 12.605 17.8711 13.28 17.0161 13.28ZM17.2111 19.565C15.8461 19.565 15.0061 19.085 15.0061 18.275C15.0061 17.855 15.2161 17.435 15.7111 17.075C16.0561 17.15 16.4311 17.195 16.8211 17.195H18.0661C19.0711 17.195 19.6111 17.405 19.6111 18.095C19.6111 18.86 18.6511 19.565 17.2111 19.565Z" />
   </svg>
+)
 
-export const SmallBold = ({size=24, className}) =>
-  <svg className={className} width={size} height={size} viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-    <path d="M8.25898 12.219L8.40298 11.652C8.57398 10.986 8.75398 10.203 8.91598 9.501H8.95198C9.13198 10.185 9.30298 10.986 9.48298 11.652L9.62698 12.219H8.25898ZM10.32 15H12.012L9.93298 8.304H8.00698L5.92798 15H7.56598L7.95298 13.461H9.93298L10.32 15Z"/>
-    <path d="M15.3254 17.151C17.1074 17.151 18.2234 16.386 18.2234 15.306C18.2234 14.379 17.5124 13.992 16.2524 13.992H15.4424C14.8934 13.992 14.6864 13.884 14.6864 13.65C14.6864 13.479 14.7314 13.398 14.8484 13.29C15.0734 13.353 15.2804 13.38 15.4514 13.38C16.5854 13.38 17.4944 12.858 17.4944 11.661C17.4944 11.418 17.4314 11.184 17.3504 11.04H18.1424V9.879H16.2614C16.0274 9.798 15.7484 9.753 15.4514 9.753C14.3444 9.753 13.3094 10.365 13.3094 11.616C13.3094 12.21 13.6424 12.687 14.0024 12.939V12.975C13.6694 13.209 13.4354 13.569 13.4354 13.929C13.4354 14.361 13.6334 14.631 13.9034 14.811V14.856C13.4264 15.099 13.1744 15.441 13.1744 15.864C13.1744 16.782 14.1464 17.151 15.3254 17.151ZM15.4514 12.417C15.0824 12.417 14.8034 12.147 14.8034 11.616C14.8034 11.094 15.0824 10.824 15.4514 10.824C15.8294 10.824 16.0994 11.094 16.0994 11.616C16.0994 12.147 15.8294 12.417 15.4514 12.417ZM15.5774 16.152C14.9204 16.152 14.4884 15.972 14.4884 15.612C14.4884 15.441 14.5694 15.297 14.7314 15.153C14.8934 15.198 15.0914 15.216 15.4604 15.216H15.9464C16.4234 15.216 16.6934 15.279 16.6934 15.585C16.6934 15.909 16.2344 16.152 15.5774 16.152Z"/>
+export const SmallBold = ({size = 24, className}) => (
+  <svg
+    className={className}
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M8.25898 12.219L8.40298 11.652C8.57398 10.986 8.75398 10.203 8.91598 9.501H8.95198C9.13198 10.185 9.30298 10.986 9.48298 11.652L9.62698 12.219H8.25898ZM10.32 15H12.012L9.93298 8.304H8.00698L5.92798 15H7.56598L7.95298 13.461H9.93298L10.32 15Z" />
+    <path d="M15.3254 17.151C17.1074 17.151 18.2234 16.386 18.2234 15.306C18.2234 14.379 17.5124 13.992 16.2524 13.992H15.4424C14.8934 13.992 14.6864 13.884 14.6864 13.65C14.6864 13.479 14.7314 13.398 14.8484 13.29C15.0734 13.353 15.2804 13.38 15.4514 13.38C16.5854 13.38 17.4944 12.858 17.4944 11.661C17.4944 11.418 17.4314 11.184 17.3504 11.04H18.1424V9.879H16.2614C16.0274 9.798 15.7484 9.753 15.4514 9.753C14.3444 9.753 13.3094 10.365 13.3094 11.616C13.3094 12.21 13.6424 12.687 14.0024 12.939V12.975C13.6694 13.209 13.4354 13.569 13.4354 13.929C13.4354 14.361 13.6334 14.631 13.9034 14.811V14.856C13.4264 15.099 13.1744 15.441 13.1744 15.864C13.1744 16.782 14.1464 17.151 15.3254 17.151ZM15.4514 12.417C15.0824 12.417 14.8034 12.147 14.8034 11.616C14.8034 11.094 15.0824 10.824 15.4514 10.824C15.8294 10.824 16.0994 11.094 16.0994 11.616C16.0994 12.147 15.8294 12.417 15.4514 12.417ZM15.5774 16.152C14.9204 16.152 14.4884 15.972 14.4884 15.612C14.4884 15.441 14.5694 15.297 14.7314 15.153C14.8934 15.198 15.0914 15.216 15.4604 15.216H15.9464C16.4234 15.216 16.6934 15.279 16.6934 15.585C16.6934 15.909 16.2344 16.152 15.5774 16.152Z" />
   </svg>
+)
 
-export const NormalBold = ({size=24, className}) =>
-  <svg className={className} width={size} height={size} viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-    <path d="M7.01205 13.292L7.20405 12.536C7.43205 11.648 7.67205 10.604 7.88805 9.668H7.93605C8.17605 10.58 8.40405 11.648 8.64405 12.536L8.83605 13.292H7.01205ZM9.76005 17H12.0161L9.24405 8.072H6.67605L3.90405 17H6.08805L6.60405 14.948H9.24405L9.76005 17Z"/>
-    <path d="M16.4339 19.868C18.8099 19.868 20.2979 18.848 20.2979 17.408C20.2979 16.172 19.3499 15.656 17.6699 15.656H16.5899C15.8579 15.656 15.5819 15.512 15.5819 15.2C15.5819 14.972 15.6419 14.864 15.7979 14.72C16.0979 14.804 16.3739 14.84 16.6019 14.84C18.1139 14.84 19.3259 14.144 19.3259 12.548C19.3259 12.224 19.2419 11.912 19.1339 11.72H20.1899V10.172H17.6819C17.3699 10.064 16.9979 10.004 16.6019 10.004C15.1259 10.004 13.7459 10.82 13.7459 12.488C13.7459 13.28 14.1899 13.916 14.6699 14.252V14.3C14.2259 14.612 13.9139 15.092 13.9139 15.572C13.9139 16.148 14.1779 16.508 14.5379 16.748V16.808C13.9019 17.132 13.5659 17.588 13.5659 18.152C13.5659 19.376 14.8619 19.868 16.4339 19.868ZM16.6019 13.556C16.1099 13.556 15.7379 13.196 15.7379 12.488C15.7379 11.792 16.1099 11.432 16.6019 11.432C17.1059 11.432 17.4659 11.792 17.4659 12.488C17.4659 13.196 17.1059 13.556 16.6019 13.556ZM16.7699 18.536C15.8939 18.536 15.3179 18.296 15.3179 17.816C15.3179 17.588 15.4259 17.396 15.6419 17.204C15.8579 17.264 16.1219 17.288 16.6139 17.288H17.2619C17.8979 17.288 18.2579 17.372 18.2579 17.78C18.2579 18.212 17.6459 18.536 16.7699 18.536Z"/>
+export const NormalBold = ({size = 24, className}) => (
+  <svg
+    className={className}
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M7.01205 13.292L7.20405 12.536C7.43205 11.648 7.67205 10.604 7.88805 9.668H7.93605C8.17605 10.58 8.40405 11.648 8.64405 12.536L8.83605 13.292H7.01205ZM9.76005 17H12.0161L9.24405 8.072H6.67605L3.90405 17H6.08805L6.60405 14.948H9.24405L9.76005 17Z" />
+    <path d="M16.4339 19.868C18.8099 19.868 20.2979 18.848 20.2979 17.408C20.2979 16.172 19.3499 15.656 17.6699 15.656H16.5899C15.8579 15.656 15.5819 15.512 15.5819 15.2C15.5819 14.972 15.6419 14.864 15.7979 14.72C16.0979 14.804 16.3739 14.84 16.6019 14.84C18.1139 14.84 19.3259 14.144 19.3259 12.548C19.3259 12.224 19.2419 11.912 19.1339 11.72H20.1899V10.172H17.6819C17.3699 10.064 16.9979 10.004 16.6019 10.004C15.1259 10.004 13.7459 10.82 13.7459 12.488C13.7459 13.28 14.1899 13.916 14.6699 14.252V14.3C14.2259 14.612 13.9139 15.092 13.9139 15.572C13.9139 16.148 14.1779 16.508 14.5379 16.748V16.808C13.9019 17.132 13.5659 17.588 13.5659 18.152C13.5659 19.376 14.8619 19.868 16.4339 19.868ZM16.6019 13.556C16.1099 13.556 15.7379 13.196 15.7379 12.488C15.7379 11.792 16.1099 11.432 16.6019 11.432C17.1059 11.432 17.4659 11.792 17.4659 12.488C17.4659 13.196 17.1059 13.556 16.6019 13.556ZM16.7699 18.536C15.8939 18.536 15.3179 18.296 15.3179 17.816C15.3179 17.588 15.4259 17.396 15.6419 17.204C15.8579 17.264 16.1219 17.288 16.6139 17.288H17.2619C17.8979 17.288 18.2579 17.372 18.2579 17.78C18.2579 18.212 17.6459 18.536 16.7699 18.536Z" />
   </svg>
+)
 
-export const LargeBold = ({size=24, className}) =>
-  <svg className={className} width={size} height={size} viewBox="0 0 24 24" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
-    <path d="M5.76501 12.365L6.005 11.42C6.29 10.31 6.59001 9.005 6.86 7.835H6.92C7.22001 8.975 7.505 10.31 7.80501 11.42L8.04501 12.365H5.76501ZM9.20001 17H12.02L8.55501 5.84H5.34501L1.88 17H4.61L5.255 14.435H8.55501L9.20001 17Z"/>
-    <path d="M17.5423 20.585C20.5123 20.585 22.3724 19.31 22.3724 17.51C22.3724 15.965 21.1873 15.32 19.0873 15.32H17.7373C16.8223 15.32 16.4773 15.14 16.4773 14.75C16.4773 14.465 16.5523 14.33 16.7473 14.15C17.1223 14.255 17.4673 14.3 17.7523 14.3C19.6423 14.3 21.1574 13.43 21.1574 11.435C21.1574 11.03 21.0524 10.64 20.9174 10.4H22.2374V8.465H19.1024C18.7123 8.33 18.2473 8.255 17.7523 8.255C15.9073 8.255 14.1823 9.275 14.1823 11.36C14.1823 12.35 14.7373 13.145 15.3373 13.565V13.625C14.7823 14.015 14.3923 14.615 14.3923 15.215C14.3923 15.935 14.7223 16.385 15.1723 16.685V16.76C14.3773 17.165 13.9573 17.735 13.9573 18.44C13.9573 19.97 15.5773 20.585 17.5423 20.585ZM17.7523 12.695C17.1373 12.695 16.6723 12.245 16.6723 11.36C16.6723 10.49 17.1373 10.04 17.7523 10.04C18.3823 10.04 18.8323 10.49 18.8323 11.36C18.8323 12.245 18.3823 12.695 17.7523 12.695ZM17.9623 18.92C16.8673 18.92 16.1473 18.62 16.1473 18.02C16.1473 17.735 16.2823 17.495 16.5523 17.255C16.8223 17.33 17.1523 17.36 17.7673 17.36H18.5774C19.3724 17.36 19.8223 17.465 19.8223 17.975C19.8223 18.515 19.0573 18.92 17.9623 18.92Z"/>
+export const LargeBold = ({size = 24, className}) => (
+  <svg
+    className={className}
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill="currentColor"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path d="M5.76501 12.365L6.005 11.42C6.29 10.31 6.59001 9.005 6.86 7.835H6.92C7.22001 8.975 7.505 10.31 7.80501 11.42L8.04501 12.365H5.76501ZM9.20001 17H12.02L8.55501 5.84H5.34501L1.88 17H4.61L5.255 14.435H8.55501L9.20001 17Z" />
+    <path d="M17.5423 20.585C20.5123 20.585 22.3724 19.31 22.3724 17.51C22.3724 15.965 21.1873 15.32 19.0873 15.32H17.7373C16.8223 15.32 16.4773 15.14 16.4773 14.75C16.4773 14.465 16.5523 14.33 16.7473 14.15C17.1223 14.255 17.4673 14.3 17.7523 14.3C19.6423 14.3 21.1574 13.43 21.1574 11.435C21.1574 11.03 21.0524 10.64 20.9174 10.4H22.2374V8.465H19.1024C18.7123 8.33 18.2473 8.255 17.7523 8.255C15.9073 8.255 14.1823 9.275 14.1823 11.36C14.1823 12.35 14.7373 13.145 15.3373 13.565V13.625C14.7823 14.015 14.3923 14.615 14.3923 15.215C14.3923 15.935 14.7223 16.385 15.1723 16.685V16.76C14.3773 17.165 13.9573 17.735 13.9573 18.44C13.9573 19.97 15.5773 20.585 17.5423 20.585ZM17.7523 12.695C17.1373 12.695 16.6723 12.245 16.6723 11.36C16.6723 10.49 17.1373 10.04 17.7523 10.04C18.3823 10.04 18.8323 10.49 18.8323 11.36C18.8323 12.245 18.3823 12.695 17.7523 12.695ZM17.9623 18.92C16.8673 18.92 16.1473 18.62 16.1473 18.02C16.1473 17.735 16.2823 17.495 16.5523 17.255C16.8223 17.33 17.1523 17.36 17.7673 17.36H18.5774C19.3724 17.36 19.8223 17.465 19.8223 17.975C19.8223 18.515 19.0573 18.92 17.9623 18.92Z" />
   </svg>
+)
