@@ -1,12 +1,12 @@
-import React from 'react'
-import { withTranslation } from 'react-i18next'
-import { withGlobalContextConsumer } from 'contexts/GlobalContext'
-import { ChevronsLeft, ChevronsRight } from 'react-feather'
 import cn from 'classnames'
+import {withGlobalContextConsumer} from 'contexts/GlobalContext'
+import React from 'react'
+import {ChevronsLeft, ChevronsRight} from 'react-feather'
+import {withTranslation} from 'react-i18next'
 import './collapse-button.scss'
 
-const CollapseButton = ({placement='left', globalSettings, changeGlobalSetting, t}) => {
-  const isLeft = placement==='left'
+const CollapseButton = ({placement = 'left', globalSettings, changeGlobalSetting, t}) => {
+  const isLeft = placement === 'left'
   const {leftCollapse, rightCollapse} = globalSettings
   const collapsed = isLeft ? leftCollapse : rightCollapse
 
@@ -20,11 +20,17 @@ const CollapseButton = ({placement='left', globalSettings, changeGlobalSetting, 
       onClick={toggleCollapse}
       title={t(`hide ${placement} sider`)}
     >
-      {
-        collapsed ?
-        (isLeft ? <ChevronsRight size={16}/> : <ChevronsLeft size={16}/>) :
-        (isLeft ? <ChevronsLeft size={16}/> : <ChevronsRight size={16}/>)
-      }
+      {collapsed ? (
+        isLeft ? (
+          <ChevronsRight size={16} />
+        ) : (
+          <ChevronsLeft size={16} />
+        )
+      ) : isLeft ? (
+        <ChevronsLeft size={16} />
+      ) : (
+        <ChevronsRight size={16} />
+      )}
     </button>
   )
 }
