@@ -1,11 +1,7 @@
-import React from 'react'
 import Tooltip from 'rc-tooltip'
+import React from 'react'
 
 const WithTooltip = ({yes, tooltipProps, children}) =>
-  yes ?
-  <Tooltip {...tooltipProps}>
-    { children }
-  </Tooltip> :
-  children
+  yes ? <Tooltip {...tooltipProps}>{children}</Tooltip> : children
 
 export default WithTooltip
