@@ -1,5 +1,5 @@
-export { default as WithTooltip } from './WithTooltip'
-export { default as WithCopy } from './WithCopy'
-export { default as InputGroup } from './InputGroup'
-export { default as CopiableInput } from './CopiableInput'
-export { default as CollapseButton } from './CollapseButton'
+export {default as WithTooltip} from './WithTooltip'
+export {default as WithCopy} from './WithCopy'
+export {default as InputGroup} from './InputGroup'
+export {default as CopiableInput} from './CopiableInput'
+export {default as CollapseButton} from './CollapseButton'
