@@ -1,8 +1,8 @@
-export { default as common } from './common.lang.json'
-export { default as entry } from 'page/entry/entry.lang.json'
-export { default as left } from 'page/left/left.lang.json'
-export { default as right } from 'page/right/right.lang.json'
-export { default as rightItems } from 'page/right/items/items.lang.json'
-export { default as utilities } from 'components/utilities/utilities.lang.json'
-export { default as canvas } from 'page/canvas/canvas.lang.json'
-export { default as header } from 'page/header/header.lang.json'
+export {default as common} from './common.lang.json'
+export {default as entry} from 'page/entry/entry.lang.json'
+export {default as left} from 'page/left/left.lang.json'
+export {default as right} from 'page/right/right.lang.json'
+export {default as rightItems} from 'page/right/items/items.lang.json'
+export {default as utilities} from 'components/utilities/utilities.lang.json'
+export {default as canvas} from 'page/canvas/canvas.lang.json'
+export {default as header} from 'page/header/header.lang.json'
