@@ -1,8 +1,8 @@
 import React from 'react'
 
-const ProductHunt = () =>
+const ProductHunt = () => (
   <div className="entry-producthunt">
-    <a
+    {/* <a
       href="https://www.producthunt.com/posts/figma-handoff?utm_source=badge-featured&utm_medium=badge&utm_souce=badge-figma-handoff"
       target="_blank"
       rel="noopener noreferrer"
@@ -11,7 +11,8 @@ const ProductHunt = () =>
         src="https://api.producthunt.com/widgets/embed-image/v1/featured.svg?post_id=184379&theme=light"
         alt="Figma Handoff - A Figma handoff tool, generate design mark for developers. | Product Hunt Embed"
       />
-    </a>
+    </a> */}
   </div>
+)
 
 export default ProductHunt
