@@ -1,18 +1,19 @@
-import React from 'react'
 import Tooltip from 'rc-tooltip'
+import React from 'react'
 
-const Overlay = ({ children, ...props }) =>
+const Overlay = ({children, ...props}) => (
   <Tooltip
     trigger={['click']}
     overlayStyle={{width: 320}}
     align={{
-      offset: [0, -10]
+      offset: [0, -10],
     }}
     placement="bottomRight"
     transitionName="rc-tooltip-slide"
     {...props}
   >
-    { children }
+    {children}
   </Tooltip>
+)
 
 export default Overlay
