@@ -1,24 +1,27 @@
-import React, { Fragment } from 'react'
 import cn from 'classnames'
-import { withTranslation } from 'react-i18next'
+import {WithTooltip} from 'components/utilities'
+import React, {Fragment} from 'react'
+import {withTranslation} from 'react-i18next'
+import {getBackgroundImageUrl, getImageUrl} from 'utils/helper'
 import Search from './Search'
-import { WithTooltip } from 'components/utilities'
-import { getImageUrl, getBackgroundImageUrl } from 'utils/helper'
 
 class Components extends React.Component {
-  defaultComponents = this.props.components.map(c =>({...c, filterVisible: true}))
+  defaultComponents = this.props.components.map(c => ({
+    ...c,
+    filterVisible: true,
+  }))
   state = {
     searchValue: '',
     components: this.defaultComponents,
-    selectedIndex: 0
+    selectedIndex: 0,
   }
   handleComponentSelect = (index, componentId) => {
-    const { onComponentChange, mode, isMock } = this.props
+    const {onComponentChange, mode, isMock} = this.props
     const componentImageUrl = getImageUrl(this.defaultComponents[index], mode, isMock)
     this.setState({
       searchValue: '',
       components: this.defaultComponents,
-      selectedIndex: index
+      selectedIndex: index,
     })
     onComponentChange && onComponentChange(componentId, componentImageUrl)
   }
@@ -26,64 +29,63 @@ class Components extends React.Component {
     const searchValue = e.target.value
     this.setState({
       searchValue,
-      components: this.defaultComponents
-        .map(c =>({
-          ...c,
-          filterVisible: c.name.toLowerCase().indexOf(searchValue.toLowerCase())>-1
-        }))
+      components: this.defaultComponents.map(c => ({
+        ...c,
+        filterVisible: c.name.toLowerCase().indexOf(searchValue.toLowerCase()) > -1,
+      })),
     })
   }
   clearSearch = () => {
     this.setState({
       searchValue: '',
-      components: this.defaultComponents
+      components: this.defaultComponents,
     })
   }
   componentDidUpdate(prevProps) {
-    const { components } = this.props
-    const { selectedIndex } = this.state
-    if (this.props.visible && (this.props.visible !== prevProps.visible) && components.length) {
+    const {components} = this.props
+    const {selectedIndex} = this.state
+    if (this.props.visible && this.props.visible !== prevProps.visible && components.length) {
       this.handleComponentSelect(selectedIndex, components[selectedIndex].id)
     }
   }
-  render () {
-    const { visible, mode, isMock, t } = this.props
-    const { searchValue, components, selectedIndex } = this.state
+  render() {
+    const {visible, mode, isMock, t} = this.props
+    const {searchValue, components, selectedIndex} = this.state
     return (
       <Fragment>
-        <Search
-          visible={visible}
-          value={searchValue}
-          onChange={this.handleSearchChange}
-          onClear={this.clearSearch}
-        />
+        <Search visible={visible} value={searchValue} onChange={this.handleSearchChange} onClear={this.clearSearch} />
         <ul className={cn('list-container list-components', {hide: !visible})}>
-          {
-            !!components.length ?
-            components.map(
-              (component, index) =>
-                <WithTooltip
-                  key={component.id}
-                  yes={!!component.description}
-                  tooltipProps={{overlay: component.description, placement: 'right'}}
-                >                                  
-                  <li
-                    className={cn('list-item', {'list-item-hidden': !component.filterVisible, selected: index===selectedIndex})}
-                    onClick={() => this.handleComponentSelect(index, component.id)}
-                    title={component.name}
-                  >
-                    <div
-                      className="item-thumbnail"
-                      style={{
-                        backgroundImage: getBackgroundImageUrl(component, mode, isMock)
-                      }}
-                    />
-                    <span>{component.name}</span>
-                  </li>
-                </WithTooltip>
-            ) :
+          {!!components.length ? (
+            components.map((component, index) => (
+              <WithTooltip
+                key={component.id}
+                yes={!!component.description}
+                tooltipProps={{
+                  overlay: component.description,
+                  placement: 'right',
+                }}
+              >
+                <li
+                  className={cn('list-item', {
+                    'list-item-hidden': !component.filterVisible,
+                    selected: index === selectedIndex,
+                  })}
+                  onClick={() => this.handleComponentSelect(index, component.id)}
+                  title={component.name}
+                >
+                  <div
+                    className="item-thumbnail"
+                    style={{
+                      backgroundImage: getBackgroundImageUrl(component, mode, isMock),
+                    }}
+                  />
+                  <span>{component.name}</span>
+                </li>
+              </WithTooltip>
+            ))
+          ) : (
             <li className="item-empty">{t('no components')}</li>
-          }
+          )}
         </ul>
       </Fragment>
     )
