@@ -1,24 +1,24 @@
-import React from 'react'
 import cn from 'classnames'
-import { withTranslation } from 'react-i18next'
-import { withGlobalContextConsumer } from 'contexts/GlobalContext'
-import { CollapseButton } from 'components/utilities'
+import {CollapseButton} from 'components/utilities'
+import {withGlobalContextConsumer} from 'contexts/GlobalContext'
+import React from 'react'
+import {withTranslation} from 'react-i18next'
+import Components from './Components'
 import Frames from './Frames'
 import NestedFrames from './NestedFrames'
-import Components from './Components'
 import './index.scss'
 
 class LeftPanel extends React.Component {
   state = {
-    tabIndex: 0
+    tabIndex: 0,
   }
-  handleTabClick = (index) => {
-    const { tabIndex } = this.state
-    if (tabIndex===index) return
-    this.setState({ tabIndex: index })
+  handleTabClick = index => {
+    const {tabIndex} = this.state
+    if (tabIndex === index) return
+    this.setState({tabIndex: index})
   }
 
-  render () {
+  render() {
     const {
       components,
       includeComponents,
@@ -28,59 +28,58 @@ class LeftPanel extends React.Component {
       globalSettings,
       onSiderTransitionEnd,
       onFrameOrComponentChange,
-      t
+      t,
     } = this.props
-    const { leftCollapse } = globalSettings
-    const { tabIndex } = this.state
+    const {leftCollapse} = globalSettings
+    const {tabIndex} = this.state
     return (
-      <div
-        className={cn('main-left', {collapsed: leftCollapse})}
-        onTransitionEnd={onSiderTransitionEnd}
-      >
+      <div className={cn('main-left', {collapsed: leftCollapse})} onTransitionEnd={onSiderTransitionEnd}>
         <CollapseButton placement="left" />
         <div className="left-sider">
-          {
-            !!includeComponents &&
+          {!!includeComponents && (
             <ul className="left-sider-tabs">
-              <li
-                className={cn({selected: tabIndex===0})}
-                onClick={() => this.handleTabClick(0)}
-              >{t('pages')}</li>
-              <li
-                className={cn({selected: tabIndex===1})}
-                onClick={() => this.handleTabClick(1)}
-              >{t('components')}</li>
+              <li className={cn({selected: tabIndex === 0})} onClick={() => this.handleTabClick(0)}>
+                {t('pages')}
+              </li>
+              <li className={cn({selected: tabIndex === 1})} onClick={() => this.handleTabClick(1)}>
+                {t('components')}
+              </li>
             </ul>
-          }
-          <div className={cn('left-sider-list', {'without-tab': !includeComponents})}>
+          )}
+          <div
+            className={cn('left-sider-list', {
+              'without-tab': !includeComponents,
+            })}
+          >
             {
               // 此处不能使用 useNestedPages 判断，因为默认为 undefined 会进入后一个，导致出错中断
-              Array.isArray(pagedFrames) ?
-              <NestedFrames
-                visible={tabIndex===0}
-                pagedFrames={pagedFrames}
-                mode={mode}
-                isMock={isMock}
-                onSelect={onFrameOrComponentChange}
-              /> :
-              <Frames
-                visible={tabIndex===0}
-                pagedFrames={pagedFrames}
-                mode={mode}
-                isMock={isMock}
-                onFrameChange={onFrameOrComponentChange}
-              />
+              Array.isArray(pagedFrames) ? (
+                <NestedFrames
+                  visible={tabIndex === 0}
+                  pagedFrames={pagedFrames}
+                  mode={mode}
+                  isMock={isMock}
+                  onSelect={onFrameOrComponentChange}
+                />
+              ) : (
+                <Frames
+                  visible={tabIndex === 0}
+                  pagedFrames={pagedFrames}
+                  mode={mode}
+                  isMock={isMock}
+                  onFrameChange={onFrameOrComponentChange}
+                />
+              )
             }
-            {
-              !!includeComponents &&
+            {!!includeComponents && (
               <Components
-                visible={tabIndex===1}
+                visible={tabIndex === 1}
                 components={components}
                 mode={mode}
                 isMock={isMock}
                 onComponentChange={onFrameOrComponentChange}
               />
-            }
+            )}
           </div>
         </div>
       </div>
