@@ -1,10 +1,10 @@
-import React, { useState, useEffect } from 'react'
-import { withGlobalContextConsumer } from 'contexts/GlobalContext'
-import { COLOR_FORMATS, ANDROID_COLOR_FORMATS } from 'utils/const'
+import {withGlobalContextConsumer} from 'contexts/GlobalContext'
+import React, {useState, useEffect} from 'react'
+import {ANDROID_COLOR_FORMATS, COLOR_FORMATS} from 'utils/const'
 import './color-format-select.scss'
 
-const ColorFormatSelect = ({ globalSettings, changeGlobalSetting }) => {
-  const [ colorFormat, setColorFormat ] = useState(globalSettings.colorFormat || 0)
+const ColorFormatSelect = ({globalSettings, changeGlobalSetting}) => {
+  const [colorFormat, setColorFormat] = useState(globalSettings.colorFormat || 0)
   const changeColorFormat = e => {
     const value = e.target.value - 0
     setColorFormat(value)
@@ -15,15 +15,17 @@ const ColorFormatSelect = ({ globalSettings, changeGlobalSetting }) => {
     setColorFormat(globalSettings.colorFormat)
   }, [globalSettings.colorFormat])
 
-  const { platform } = globalSettings
+  const {platform} = globalSettings
 
-  return <select className="input color-format-select" value={colorFormat} onChange={changeColorFormat}>
-    {
-      (platform===2 ? ANDROID_COLOR_FORMATS : COLOR_FORMATS).map((format, index) =>
-        <option key={index} value={index}>{ format }</option>
-      )
-    }
-  </select>
+  return (
+    <select className="input color-format-select" value={colorFormat} onChange={changeColorFormat}>
+      {(platform === 2 ? ANDROID_COLOR_FORMATS : COLOR_FORMATS).map((format, index) => (
+        <option key={index} value={index}>
+          {format}
+        </option>
+      ))}
+    </select>
+  )
 }
 
 export default withGlobalContextConsumer(ColorFormatSelect)
