@@ -1,8 +1,8 @@
-export { default as FillItem } from './FillItem'
-export { default as EffectItem } from './EffectItem'
-export { default as TextItems } from './TextItems'
-export { default as ExportItem } from './ExportItem'
-export { default as StyleItem } from './StyleItem'
-export { default as ColorFormatSelect } from './ColorFormatSelect'
-export { default as Preview } from './Preview'
-export { default as Color } from './Color'
+export {default as FillItem} from './FillItem'
+export {default as EffectItem} from './EffectItem'
+export {default as TextItems} from './TextItems'
+export {default as ExportItem} from './ExportItem'
+export {default as StyleItem} from './StyleItem'
+export {default as ColorFormatSelect} from './ColorFormatSelect'
+export {default as Preview} from './Preview'
+export {default as Color} from './Color'
