@@ -1,13 +1,13 @@
-export function getAllFrames (nestedFrames) {
+export function getAllFrames(nestedFrames) {
   const frames = []
-  function walk (items, pageId) {
+  function walk(items, pageId) {
     // eslint-disable-next-line array-callback-return
     items.map(item => {
       if (!item.id.startsWith('temp-')) {
         frames.push({
           id: item.id,
           name: item.name,
-          pageId
+          pageId,
         })
       }
       if (item.children) {
@@ -21,19 +21,19 @@ export function getAllFrames (nestedFrames) {
 }
 
 export const getPagedFrames = data => {
-  let pagedFrames = {}
+  const pagedFrames = {}
   if (data.document && data.document.children) {
     data.document.children
       // eslint-disable-next-line
-      .map((page) => {
+      .map(page => {
         pagedFrames[page.id] = {
           isCollapsed: true,
           checked: true,
           name: page.name,
           frames: page.children
-            .filter(frame => frame.type==='FRAME' && frame.visible!==false)
+            .filter(frame => frame.type === 'FRAME' && frame.visible !== false)
             .map(({id, name}) => ({id, name, checked: true}))
-            .reverse()
+            .reverse(),
         }
       })
   }
@@ -41,15 +41,16 @@ export const getPagedFrames = data => {
 }
 
 export const getSelectedPagedFrames = pagedFrames => {
-  let finalFrames = {}
+  const finalFrames = {}
   Object.keys(pagedFrames)
     // eslint-disable-next-line
     .map(pageId => {
-      const { name, frames } = pagedFrames[pageId]
+      const {name, frames} = pagedFrames[pageId]
       const checkedFrames = frames.filter(({checked}) => checked)
       if (checkedFrames.length) {
         finalFrames[pageId] = {
-          name, frames: checkedFrames
+          name,
+          frames: checkedFrames,
         }
       }
     })
@@ -61,12 +62,12 @@ export const getFrameOptions = pagedFrames => {
   Object.keys(pagedFrames)
     // eslint-disable-next-line
     .map(pageId => {
-      const { name, frames } = pagedFrames[pageId]
+      const {name, frames} = pagedFrames[pageId]
       if (frames.length) {
         options.push({
           label: name,
           value: pageId,
-          children: frames.map(({id, name}) => ({value: id, label: name}))
+          children: frames.map(({id, name}) => ({value: id, label: name})),
         })
       }
     })
@@ -74,32 +75,32 @@ export const getFrameOptions = pagedFrames => {
 }
 
 export const filterFrameOptions = (frameOptions, value) => {
-  return frameOptions.map(p => {
-    const pageContainsValue = p.label.toLowerCase().indexOf(value.toLowerCase()) > -1
-    const frameContainsValue = p.children.some(f => f.label.toLowerCase().indexOf(value.toLowerCase()) > -1)
-    if (pageContainsValue) {
-      return p
-    } else if (frameContainsValue) {
-      const targetedChildren = p.children.filter(f => f.label.toLowerCase().indexOf(value.toLowerCase()) > -1)
-      return {...p, children: targetedChildren}
-    } else {
-      return ''
-    }
-  })
-  .filter(p => !!p)
+  return frameOptions
+    .map(p => {
+      const pageContainsValue = p.label.toLowerCase().indexOf(value.toLowerCase()) > -1
+      const frameContainsValue = p.children.some(f => f.label.toLowerCase().indexOf(value.toLowerCase()) > -1)
+      if (pageContainsValue) {
+        return p
+      } else if (frameContainsValue) {
+        const targetedChildren = p.children.filter(f => f.label.toLowerCase().indexOf(value.toLowerCase()) > -1)
+        return {...p, children: targetedChildren}
+      } else {
+        return ''
+      }
+    })
+    .filter(p => !!p)
 }
 
-export const getFlattenedFrames = (pagedFrames, needCheck=true) => {
+export const getFlattenedFrames = (pagedFrames, needCheck = true) => {
   let flattenedFrames = []
   Object.keys(pagedFrames)
     // eslint-disable-next-line
     .map(pageId => {
-      flattenedFrames = flattenedFrames
-        .concat(
-          pagedFrames[pageId].frames
-            .filter(({checked}) => needCheck ? checked : true)
-            .map(frame => ({...frame, pageId}))
-        )
+      flattenedFrames = flattenedFrames.concat(
+        pagedFrames[pageId].frames
+          .filter(({checked}) => (needCheck ? checked : true))
+          .map(frame => ({...frame, pageId})),
+      )
     })
   return flattenedFrames
 }
