@@ -1,31 +1,29 @@
-import { LOCALIZED_SETTING_KEYS } from "./const"
+import {LOCALIZED_SETTING_KEYS} from './const'
 
 export const throttle = (fn, delay) => {
- 	var timer = null
- 	return function(){
- 		var context = this, args = arguments
- 		clearTimeout(timer)
- 		timer = setTimeout(function(){
- 			fn.apply(context, args)
- 		}, delay)
- 	}
+  var timer = null
+  return function () {
+    var args = arguments
+    clearTimeout(timer)
+    timer = setTimeout(() => {
+      fn.apply(this, args)
+    }, delay)
+  }
 }
 
-export const isCmdOrCtrl = e => navigator.platform.match("Mac") ? e.metaKey : e.ctrlKey
+export const isCmdOrCtrl = e => (navigator.platform.match('Mac') ? e.metaKey : e.ctrlKey)
 
-export const getStyleItems = (node, key) =>
-  key==='backgrounds' ? node.fills : node[key]
+export const getStyleItems = (node, key) => (key === 'backgrounds' ? node.fills : node[key])
 
-export const isAllImageFill = fills =>
-  fills.filter(fill => fill.type === 'IMAGE').length === fills.length
+export const isAllImageFill = fills => fills.filter(fill => fill.type === 'IMAGE').length === fills.length
 
 export const getFileName = (exportSetting, index) => {
-  const { name, suffix, format, constraint } = exportSetting
+  const {name, suffix, format, constraint} = exportSetting
   let fileName = suffix ? `${name}-${suffix}` : name
-  if (index!==undefined) {
+  if (index !== undefined) {
     fileName += `-${index}`
   }
-  const scale = format==='SVG' ? '' : `@${constraint.value}x`
+  const scale = format === 'SVG' ? '' : `@${constraint.value}x`
   const fileFormat = format.toLowerCase()
   fileName = fileName.replace(/ /g, '-')
   return `${trimFilePath(fileName)}${scale}.${fileFormat}`
@@ -35,7 +33,7 @@ export const formatStyles = styles => {
   const fatStyles = {}
   // eslint-disable-next-line
   Object.keys(styles).map(key => {
-    const { styleType } = styles[key]
+    const {styleType} = styles[key]
     if (fatStyles[styleType]) {
       fatStyles[styleType].push({id: key, ...styles[key]})
     } else {
@@ -51,14 +49,12 @@ export async function asyncForEach(array, callback) {
   }
 }
 
-export function getImageUrl (item, mode, isMock) {
+export function getImageUrl(item, mode, isMock) {
   const fileName = item.fileName ? `exports/${item.fileName}` : `${item.id.replace(/:/g, '-')}.png`
-  return (mode==='local' || isMock) ?
-    `${process.env.PUBLIC_URL}/data/${fileName}` :
-    item.image.url
+  return mode === 'local' || isMock ? `${process.env.PUBLIC_URL}/data/${fileName}` : item.image.url
 }
 
-export function getBackgroundImageUrl (item, mode, isMock) {
+export function getBackgroundImageUrl(item, mode, isMock) {
   return `url(${getImageUrl(item, mode, isMock)})`
 }
 
@@ -73,15 +69,13 @@ export const onInputClick = (e, callback) => {
 
 export const copySomething = (text, callback) => {
   return () => {
-    navigator.clipboard.writeText(text)
-      .then(callback)
+    navigator.clipboard.writeText(text).then(callback)
   }
 }
 
-export const trimFilePath = filePath =>
-  filePath.replace(/\//g, '-').replace(/:/g, '-')
+export const trimFilePath = filePath => filePath.replace(/\//g, '-').replace(/:/g, '-')
 
-export const filterLocalizedSettings = (settings) => {
+export const filterLocalizedSettings = settings => {
   const filteredSettings = {}
   // eslint-disable-next-line
   Object.keys(settings).map(key => {
@@ -93,9 +87,9 @@ export const filterLocalizedSettings = (settings) => {
 }
 
 export const getLocalGlobalSettings = () => {
-  return JSON.parse(window.localStorage.getItem('heronHandoff.settings'))
+  return JSON.parse(window.localStorage.getItem('f2cHandoff.settings'))
 }
 
-export const setLocalGlobalSettings = (settings) => {
-  window.localStorage.setItem('heronHandoff.settings', JSON.stringify(filterLocalizedSettings(settings)))
+export const setLocalGlobalSettings = settings => {
+  window.localStorage.setItem('f2cHandoff.settings', JSON.stringify(filterLocalizedSettings(settings)))
 }
